requirements.txt
adjustText==1.0.4
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.7
Django==4.2.7
django-cookies-samesite==0.9.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.13.1
fonttools==4.49.0
gunicorn==21.2.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pipenv==2023.10.24
platformdirs==3.11.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
ua-parser==0.10.0
urllib3==2.2.1
virtualenv==20.24.6
window==0.0.3